+++ ui/src/wt/mod.rs
@@ -27,8 +27,9 @@ pub use view_selected::{WtSelectedLayerView, WtSelectedLayerResponse};
 pub use morph::{morph_amount_for_position, morph_position, WtMorph, WtMorphResponse};
 pub use quant_handles::{
     apply_quant_slot_amplitude, frame_to_y, knob_y_on_curve, nearest_quant_handle, nearest_slot,
-    paint_quant_knob, quant_control_points, quant_curve_stroke, quant_hover_status_label,
-    quant_knob_visual, quantized_curve_polyline, resample_frame_from_quant_points, resample_frame_from_quant_points_uniform,
+    paint_quant_knob, periodize_quant_frame, quant_control_points, quant_curve_stroke,
+    quant_hover_status_label, quant_knob_visual, quantized_curve_polyline,
+    resample_frame_from_quant_points, resample_frame_from_quant_points_uniform,
     sample_at_quant_phase, sample_from_knob_y, sample_to_y, snap_x_to_slot, slot_x, y_to_frame,
     y_to_sample, QuantHandleEditor, QuantHandleResponse, QuantKnobVisual,
 };
@@ -43,4 +44,6 @@ pub use toolbar::{FrameShapeTemplate, WtEditTool, WtToolbar, WtToolbarResponse};
 pub use view_2d::{apply_frame_shape_template, shape_template_source_type, WtView2d, WtView2dResponse};
 pub use view_3d::{WtView3d, WtView3dResponse};
 pub use view_3d_stack::{composite_stack_sample, WtView3dStack, WtView3dStackResponse};
-pub use waveform::{frame_index, waveform_points};
+pub use waveform::{
+    frame_index, layer_quant_editable, quant_knobs_for_selection, waveform_points,
+};